% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abra}
\alias{abra}
\title{Subset of Abra alba data from OBIS.}
\format{An object of class \code{data.frame} with 2093 rows and 44 columns.}
\usage{
abra
}
\description{
Subset of Abra alba data from OBIS.
}
\keyword{datasets}
