% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_centroid.R
\name{calculate_centroid}
\alias{calculate_centroid}
\title{Calculates the centroid and radius of a WKT geometry.}
\usage{
calculate_centroid(wkt)
}
\arguments{
\item{wkt}{character vector One or more WKT strings with longitude/latitude
coordinates (epsg=4326)}
}
\value{
Data frame with centroid coordinates and radius in meter.
}
\description{
Calculates the centroid and radius of a WKT geometry.
}
