% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_depth.R
\name{check_depth}
\alias{check_depth}
\title{Check which points are located on land.}
\usage{
check_depth(data, report = FALSE)
}
\arguments{
\item{data}{The data frame.}

\item{depthmargin}{How much can the given depth deviate from the bathymetry
in the rasters (in meters).}

\item{shoremargin}{How far offshore should a record be to have a bathymetry
larger then 0. If \code{NA} (default) then this test is ignored}

\item{report}{If TRUE, errors are returned instead of records.}
}
\value{
Errors or records.
}
\description{
Check which points are located on land.
}
