% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eventdate.R
\name{check_eventdate}
\alias{check_eventdate}
\title{Check eventDate format.}
\usage{
check_eventdate(data)
}
\arguments{
\item{data}{The data frame.}
}
\value{
Any errors.
}
\description{
Check eventDate format.
}
