% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eventids.R
\name{check_eventids}
\alias{check_eventids}
\title{Check if eventID and parentEventID are present, and parentEventIDs have corresponding eventIDs.}
\usage{
check_eventids(event)
}
\arguments{
\item{event}{The event records.}
}
\value{
Any errors.
}
\description{
Check if eventID and parentEventID are present, and parentEventIDs have corresponding eventIDs.
}
