% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eventids.R
\name{check_extension_eventids}
\alias{check_extension_eventids}
\title{Check if all eventIDs in an extension have corresponding eventIDs in the core.}
\usage{
check_extension_eventids(event, extension)
}
\arguments{
\item{event}{The event records.}

\item{event}{The extension records.}
}
\value{
Any errors.
}
\description{
Check if all eventIDs in an extension have corresponding eventIDs in the core.
}

