% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_onland.R
\name{check_onland}
\alias{check_onland}
\title{Check which points are located on land.}
\usage{
check_onland(data, land = NULL, report = FALSE, buffer = 0,
  offline = FALSE)
}
\arguments{
\item{data}{The data frame.}

\item{land}{SpatialPolygonsDataFrame. If not provided the simplified land
polygons from OSM are used. This parameter is ignored when, \code{offline =
FALSE}.}

\item{report}{If TRUE, errors are returned instead of records.}

\item{buffer}{Set how far inland points are still to be deemed valid (in meters).}

\item{offline}{If TRUE, a local simplified shoreline is used, otherwise an
OBIS webservice is used. The default value is \code{FALSE}.}
}
\value{
Errors or records.
}
\description{
Check which points are located on land.
}
\examples{
report <- check_onland(abra, report = TRUE, buffer = 100)
print(report)
# plot records on land with 100 meter buffer
plot_map_leaflet(abra[report$row,], popup = "id")
# filter records not on land
abra[-1 * report$row,]
check_onland(abra, report = FALSE, buffer = 100)
}
