% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_outliers_species}
\alias{check_outliers_species}
\title{Check which points are spatial dataset outliers.}
\usage{
check_outliers_species(data, report = FALSE, mad_coef = 6, iqr_coef =
  3)
}
\arguments{
\item{data}{The data frame with decimalLongitude, decimalLatitude and
scientificNameID columns.}

\item{report}{If TRUE, errors are returned instead of records.}

\item{mad_coef}{Coefficient to multiply the median absolute deviation (MAD)
by in order to determine the range of valid values. Default is \code{6}.}

\item{iqr_coef}{Coefficient to multiply the interquartile range (IQR) by in
order to determine the range of valid values. Default values is \code{3}.}

\item{topn}{Number of species for which the QC should be performed. Default
is \code{NA} (all species).}
}
\value{
Problematic records or an errors report.
}
\description{
Check which points are spatial dataset outliers.
}
\examples{
\dontrun{
notok <- check_outliers_species(abra, report = FALSE)
print(nrow(notok))
r <- check_outliers_species(abra, report = TRUE, mad_coef = 3, iqr_coef = 1.5)
print(r)
plot_map_leaflet(abra[r$row,], popup = "id")
plot_outliers(r)
}
}
\seealso{
\code{\link{plot_outliers}} \code{\link{check_outliers_dataset}}
  \code{\link{check_onland}} \code{\link{check_depth}}
}
