% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{exportTree}
\alias{exportTree}
\title{Exports a visual representation of a tree to an HTML file.}
\usage{
exportTree(tree, filename)
}
\description{
Exports a visual representation of a tree to an HTML file.
}
