% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_event}
\alias{flatten_event}
\title{Flatten event records.}
\usage{
flatten_event(event, fields = NULL)
}
\arguments{
\item{event}{The event records.}

\item{fields}{Fields to be inherited from higher levels, if NULL all Event Core fields will be inherited.}
}
\value{
Flattened event records.
}
\description{
Flatten event records.
}
