% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_occurrence}
\alias{flatten_occurrence}
\title{Flatten event and occurrence records.}
\usage{
flatten_occurrence(event, occurrence, fields = NULL)
}
\arguments{
\item{event}{The event records.}

\item{fields}{Fields to be inherited from the events, if NULL all Event Core fields will be inherited.}

\item{event}{The occurrence records.}
}
\value{
Flattened occurrence records.
}
\description{
Flatten event and occurrence records.
}
