% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hyperbenthos}
\alias{hyperbenthos}
\title{Hyperbenthic communities of the North Sea.}
\format{An object of class \code{dwca_gbif} of length 5.}
\usage{
hyperbenthos
}
\description{
Hyperbenthic communities of the North Sea.
}
\keyword{datasets}
