% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{identify_map}
\alias{identify_map}
\title{Deprecated: Identify a point on a map.}
\usage{
identify_map(data)
}
\arguments{
\item{data}{Original data that was plotted on the map.}
}
\value{
The nearest record.
}
\description{
Deprecated: Identify a point on a map.
}
\examples{
\dontrun{
plot_map(abra, zoom = TRUE)
identify_map(abra)
}
}
