% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land}
\alias{land}
\title{Land polygons used in the offline version of check_onland}
\format{An object of class \code{SpatialPolygonsDataFrame} with 62622 rows and 1 columns.}
\usage{
land
}
\description{
Land polygons used in the offline version of check_onland
}
\keyword{datasets}
