% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manuela}
\alias{manuela}
\title{MANUELA dataset.}
\format{An object of class \code{data.frame} with 702 rows and 9 columns.}
\usage{
manuela
}
\description{
MANUELA dataset.
}
\keyword{datasets}

