% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fields.R
\name{map_fields}
\alias{map_fields}
\title{Map column names.}
\usage{
map_fields(data, mapping)
}
\arguments{
\item{data}{The data frame.}

\item{mapping}{The mapping as a list.}
}
\value{
The data frame with mapped column names.
}
\description{
Map column names.
}
\examples{
map_fields(data.frame(x=1:3, y=4:6),
           list(decimalLongitude="x", decimalLatitude="y"))
}
