% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_taxa.R
\name{match_taxa}
\alias{match_taxa}
\title{Taxon matching using WoRMS.}
\usage{
match_taxa(names, ask = TRUE)
}
\arguments{
\item{names}{Vector of scientific names.}

\item{ask}{Ask user in case of multiple matches.}
}
\value{
Data frame with scientific name, scientific name ID and match type.
}
\description{
Taxon matching using WoRMS.
}
