% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obistools.R
\docType{package}
\name{obistools}
\alias{obistools}
\alias{obistools-package}
\title{obistools: Tools for data enhancement and quality control}
\description{
Tools for data enhancement and quality control for marine species observations. Commonly used to prepare data sets for submission to the Ocean Biogeographical Information System (<http://iobis.org>).
}
