% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot occurrences on a map.}
\usage{
plot_map(data, zoom = FALSE)
}
\arguments{
\item{data}{The data frame.}

\item{zoom}{Zoom to the occurrences (default: \code{FALSE}).}
}
\description{
Plot occurrences on a map.
}
\examples{
plot_map(abra)
plot_map(abra, zoom = TRUE)
}
