% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map_leaflet}
\alias{plot_map_leaflet}
\title{Create a Leaflet map.}
\usage{
plot_map_leaflet(data)
}
\arguments{
\item{data}{The data frame.}
}
\value{
HTML widget object.
}
\description{
Create a Leaflet map.
}

