% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Plot spatial and environmental outliers}
\usage{
plot_outliers(report)
}
\arguments{
\item{report}{The report generated by \code{\link{check_outliers_species}} or
\code{\link{check_outliers_dataset}}}
}
\description{
Plot spatial and environmental outliers
}
\details{
Note that this might generate multiple visualizations in the Plots
  and Viewers panes in RStudio.
}
\examples{
\dontrun{
r <- check_outliers_species(abra, report = TRUE)
plot_outliers(r)
r <- check_outliers_dataset(abra, report = TRUE)
plot_outliers(r)
}
}
\seealso{
\code{\link{check_outliers_species}}
  \code{\link{check_outliers_dataset}}
}
