% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Creates a basic data quality report.}
\usage{
report(data, qc = NULL, file = "report.html")
}
\arguments{
\item{data}{The data frame.}

\item{qc}{QC errors, if not provided some tests are done on the provided data.}

\item{file}{Output file.}
}
\description{
Creates a basic data quality report.
}

