% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_centroid.R
\name{calculate_centroid}
\alias{calculate_centroid}
\title{Calculates the centroid and radius of a WKT geometry.}
\usage{
calculate_centroid(wkt)
}
\arguments{
\item{wkt}{character vector One or more WKT strings with longitude/latitude
coordinates (epsg=4326)}
}
\value{
Data frame with centroid coordinates and radius in meter.
}
\description{
Calculates the centroid and radius of a WKT geometry.
}
\details{
The centroid is defined as the centroid of the convex hull of the
  provided geometry.
}
\examples{
calculate_centroid("POLYGON ((-1 -1, -1 1, 1 1, 1 -1, -1 -1))")
calculate_centroid("MULTIPOLYGON (((-1 -1, -1 1, 1 1, 1 -1, -1 -1)))")
calculate_centroid("LINESTRING (-1 -1, 0 -1, 0 1, 1 1)")
calculate_centroid("MULTILINESTRING ((-1 -1, 0 -1, 0 1, 1 1), (-3 -3, 3 3))")
calculate_centroid("POINT (0 0)")
calculate_centroid("MULTIPOINT ((0 1), (0 -1))")
}
