% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_onland.R
\name{check_onland}
\alias{check_onland}
\title{Check which points are located on land.}
\usage{
check_onland(data, polygons = NULL, report = FALSE)
}
\arguments{
\item{data}{The data frame.}

\item{polygons}{SpatialPolygonsDataFrame. If not provided the simplified land polygons from OSM are used.}

\item{report}{If TRUE, errors are returned instead of records}
}
\value{
Errors or records.
}
\description{
Check which points are located on land.
}
