% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{exportTree}
\alias{exportTree}
\title{Exports a visual representation of a tree to an HTML file.}
\usage{
exportTree(tree, filename, view = TRUE)
}
\arguments{
\item{tree}{tree as generated by \code{\link{treeStructure}}}

\item{filename}{output html file name}

\item{view}{logical. If \code{TRUE} (default) then the tree structure is opened in a
browser}
}
\description{
Exports a visual representation of a tree to an HTML file.
}
