% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_occurrence}
\alias{flatten_occurrence}
\title{Flatten event and occurrence records.}
\usage{
flatten_occurrence(event, occurrence, field = "eventID", fields = NULL)
}
\arguments{
\item{event}{The event records.}

\item{occurrence}{The occurrence records.}

\item{field}{The eventID field name in the extension records.}

\item{fields}{Fields to be inherited from the events, if NULL all Event Core fields will be inherited.}
}
\value{
Flattened occurrence records.
}
\description{
Flatten event and occurrence records.
}
