% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xylookup.R
\name{lookup_xy}
\alias{lookup_xy}
\title{Lookup spatial data for a set of points.}
\usage{
lookup_xy(data, shoredistance=TRUE, grids=TRUE, areas=FALSE,
  asdataframe=TRUE)
}
\arguments{
\item{data}{The data frame with columns decimalLongitude and decimalLatitude.}

\item{shoredistance}{Indicate whether the shoredistance should be returned
(default \code{TRUE}).}

\item{grids}{Indicate whether the grid values such as temperature and
bathymetry should be returned (default \code{TRUE})}

\item{areas}{Indicate whether the area values should be returned and from
which distance in meters to the provided points (default \code{FALSE}).}

\item{asdataframe}{Indicate whether a dataframe or a list should be returned
(default \code{TRUE}).}
}
\value{
Data frame or list with the values for the different requested
  fields.
}
\description{
Lookup spatial data for a set of points.
}
\details{
When \code{asdataframe} is \code{FALSE} then data is returned in the
  same order as the requested data as a list, with for each list item the
  requested values. For invalid coordinates \code{NULL} is returned.

  When parameter \code{areas} is a positive integer then all areas within a radius of
  this distance in meters will be returned. A value of \code{TRUE} is
  equivalent to a distance of 0 meters, \code{FALSE} indicates that no area
  results are required.
}
\examples{
\dontrun{
lookup_xy(abra, shoredistance = TRUE, grids = TRUE, areas = FALSE)
}
}
\seealso{
\code{\link{check_onland}} \code{\link{check_depth}}
}
