% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Creates a basic data quality report.}
\usage{
report(data, qc = NULL, file = "report.html", dir = NULL, view = TRUE,
  topnspecies = 20)
}
\arguments{
\item{data}{The data frame.}

\item{qc}{QC errors, if not provided some tests are done on the provided
data.}

\item{file}{Output file (default is "report.html").}

\item{dir}{Directory where to store the file (default is
\code{rappdirs::user_cache_dir("obistools")}).}

\item{view}{Logical, show the report in a browser after creation (default
\code{TRUE}).}

\item{topnspecies}{Integer, number of species ordered by number of records
for which you want to do the outlier analysis}
}
\value{
Returns the full path to the generated html report.
}
\description{
Creates a basic data quality report.
}
\examples{
\dontrun{
report(abra)
}
}
