% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{treeStructure}
\alias{treeStructure}
\title{Generate a tree summarizing the dataset structure, based on measurement types and event remarks.}
\usage{
treeStructure(event, measurement)
}
\description{
Generate a tree summarizing the dataset structure, based on measurement types and event remarks.
}
